+++ api/client/user/index.js
@@ -7,6 +7,7 @@ router.post('/login', require('./post.login'));
 router.post('/code', require('./post.create-verify-code'));
 router.post('/validate-code', require('./post.validate-code'));
 router.post('/reset-code', require('./post.create-reset-password-verify-code'));
+router.post('/validate-reset-code', require('./post.validate-reset-password-code'));
 
 router.put('/reset-password', require('./put.reset-password'));